requirements.txt
asgiref==3.8.1
Django==5.0.6
django-debug-toolbar==4.4.5
djangorestframework==3.15.2
flake8==7.0.0
flake8-docstrings==1.7.0
mccabe==0.7.0
pep8-naming==0.14.1
pycodestyle==2.11.1
pydocstyle==6.3.0
pyflakes==3.2.0
snowballstemmer==2.2.0
sqlparse==0.5.0
